# Reducing read_metadata output size in pyarrow/parquet