--- a/dask/dataframe/io/tests/test_parquet.py
+++ b/dask/dataframe/io/tests/test_parquet.py
@@ -1968,9 +1968,7 @@ def test_timeseries_nulls_in_schema(tmpdir, engine):
         )
 
 
-def test_graph_size_pyarrow(tmpdir):
-    check_pyarrow()
-
+def test_graph_size_pyarrow(tmpdir, engine):
     import pickle
 
     fn = str(tmpdir)
@@ -1979,8 +1977,7 @@ def test_graph_size_pyarrow(tmpdir):
         start="2000-01-01", end="2000-01-02", freq="60S", partition_freq="1H"
     )
 
-    eng = "pyarrow"
-    ddf1.to_parquet(fn, engine=eng)
-    ddf2 = dd.read_parquet(fn, engine=eng)
+    ddf1.to_parquet(fn, engine=engine)
+    ddf2 = dd.read_parquet(fn, engine=engine)
 
     assert len(pickle.dumps(ddf2.__dask_graph__())) < 10000