This is possible fix for the large-dask-graph problem raised in #5357.  We avoid passing around a `pq.ParquetDataset` "piece", and instead pass around the `path` and `partition_keys` members of each `piece`.  This dramitically reduces the amount of metadata stored in the task graph.

Since this solution does require the metadata to be parsed by each task in `read_partition`, I am including the results of a simple benchmark (performed on my local machine):

```python
import dask
import dask.dataframe as dd
import cloudpickle

ddf = dask.datasets.timeseries(...)
ddf.to_parquet(file, engine="pyarrow")

read_df = dd.read_parquet(file, engine="pyarrow")
fun = read_df.__dask_graph__().values()[0]
print("Graph Size:", len(cloudpickle.dumps(fun)))
%timeit read_df.compute()
```

**OUTPUT**...

*OLD ~365 partitions*:
```
Graph Size: 206328
5.71 s ± 22.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

*NEW ~365 partitions*:
```
Graph Size: 2022
5.74 s ± 101 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

*OLD ~36k partitions*:
```
Graph Size: 2144624
16.5 s ± 648 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

*NEW ~36k partitions*:
```
Graph Size: 2022
16 s ± 1.24 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```




- [x] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`