@mrocklin 2019-09-10
> Since this solution does require the metadata to be parsed by each task in read_partition, I am including the results of a simple benchmark (performed on my local machine):

@rjzamora if you're willing to humor me, can you try one of these benchmarks also with the distributed scheduler?  To do this you would add the following line of code sometime before you call `.compute()`

```python
from dask.distributed import Client
client = Client()
```