@mrocklin 2019-09-10
I would expect this to accentuate the slowdown of moving large graphs around.