@mrocklin 2019-09-10
@birdsarah , this might help out some of your workloads (if you're still interested in engaging here).  I suspect that this might have been slowing you down on many-partition workloads.