@rjzamora 2019-09-10
> if you're willing to humor me, can you try one of these benchmarks also with the distributed scheduler? To do this you would add the following line of code sometime before you call .compute()

Absolutely - Sorry, should have done this the first time around :)