@rjzamora 2019-09-10
`dask.distributed` version of [the benchmark above](https://github.com/dask/dask/pull/5391#issue-316159393)...

*OLD ~365 partitions*:
```
Graph Size: 206328
9.63 s ± 139 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

*NEW ~365 partitions*:
```
Graph Size: 2022
8.54 s ± 191 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

For `dask.distributed`, the ~36k-partition example takes **~13.6s** for the new code, but does not work for the "old" code.  I get the following output:

<details>
<summary>Click to Expand</summary>
<br>
```
/Users/rzamora/.local/lib/python3.7/site-packages/distributed/worker.py:2794: UserWarning: Large object of size 2.14 MB detected in task graph: 
  (<function ArrowEngine.read_partition at 0x11c37ea ... gories': None})
Consider scattering large objects ahead of time
with client.scatter to reduce scheduler burden and 
keep data on workers

    future = client.submit(func, big_data)    # bad

    big_future = client.scatter(big_data)     # good
    future = client.submit(func, big_future)  # good
  % (format_bytes(len(b)), s))
distributed.nanny - WARNING - Worker exceeded 95% memory budget. Restarting
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/__init__.py", line 1170, in memory_info
    return self._proc.memory_info()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 469, in memory_info
    rawtuple = self._get_pidtaskinfo()
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 339, in wrapper
    return fun(self, *args, **kwargs)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 345, in wrapper
    return fun(self)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/contextlib.py", line 130, in __exit__
    self.gen.throw(type, value, traceback)
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 372, in catch_zombie
    raise AccessDenied(proc.pid, proc._name)
psutil.AccessDenied: psutil.AccessDenied (pid=6545)
tornado.application - ERROR - Exception in callback <bound method Nanny.memory_monitor of <Nanny: tcp://127.0.0.1:56387, threads: 3>>
Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: 'Process' object has no attribute '_cache'

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_common.py", line 342, in wrapper
    ret = self._cache[fun]
AttributeError: _cache

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 359, in catch_zombie
    yield
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/psutil/_psosx.py", line 400, in _get_pidtaskinfo
    ret = cext.proc_pidtaskinfo_oneshot(self.pid)
ProcessLookupError: [Errno 3] No such process

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/anaconda3/envs/dask-parquet-dev/lib/python3.7/site-packages/tornado/ioloop.py", line 907, in _run
    return self.callback()
  File "/Users/rzamora/.local/lib/python3.7/site-packages/distributed/nanny.py", line 266, in memory_monitor
    memory = proc.memory_info().rss
…